Fix Build
Fix issue with not including the proper packages in the pip module causing it to fail the build upstream but pass locally when running dev mode since all folders in your local directory are included in dev mode but not in prod. Specifying packages in the setup.py file fixed this.