Add in logging and fix base_path bug
This commit fixes a bug for when the included file is in the current directory because I was not handling it properly before. It also adds in logging information so you can tell if the script actual ran or not instead of running and just exiting silently likely leaving users confused.